+++ lib/src/room.dart
@@ -187,7 +187,8 @@ class Room {
         state.relationshipEventId != null &&
         state.relationshipType == RelationshipTypes.edit &&
         lastEvent != null &&
-        !lastEvent.matchesEventOrTransactionId(state.relationshipEventId)) {
+        !state.matchesEventOrTransactionId(lastEvent.eventId) &&
+        lastEvent.eventId != state.relationshipEventId) {
       return;
     }
 
@@ -202,7 +203,8 @@ class Room {
     final prevEvent = getState(state.type, stateKey);
     if (prevEvent != null &&
         prevEvent.eventId != state.eventId &&
-        client.database?.eventIsKnown(state.eventId, roomId) == true) {
+        prevEvent.originServerTs.millisecondsSinceEpoch >
+            state.originServerTs.millisecondsSinceEpoch) {
       return;
     }
 
@@ -1354,15 +1356,18 @@ class Room {
   }
 
   /// Uploads a new user avatar for this room. Returns the event ID of the new
-  /// m.room.avatar event.
-  Future<String> setAvatar(MatrixFile file) async {
-    final uploadResp =
-        await client.uploadContent(file.bytes, filename: file.name);
+  /// m.room.avatar event. Leave empty to remove the current avatar.
+  Future<String> setAvatar(MatrixFile? file) async {
+    final uploadResp = file == null
+        ? null
+        : await client.uploadContent(file.bytes, filename: file.name);
     return await client.setRoomStateWithKey(
       id,
       EventTypes.RoomAvatar,
       '',
-      {'url': uploadResp.toString()},
+      {
+        if (uploadResp != null) 'url': uploadResp.toString(),
+      },
     );
   }