+++ test/commands_test.dart
@@ -274,6 +274,28 @@ void main() {
       });
     });
 
+    test('dm', () async {
+      FakeMatrixApi.calledEndpoints.clear();
+      await room.sendTextEvent('/dm @alice:example.com --no-encryption');
+      expect(
+          json.decode(
+              FakeMatrixApi.calledEndpoints['/client/r0/createRoom']?.first),
+          {
+            'invite': ['@alice:example.com'],
+            'is_direct': true,
+            'preset': 'trusted_private_chat'
+          });
+    });
+
+    test('create', () async {
+      FakeMatrixApi.calledEndpoints.clear();
+      await room.sendTextEvent('/create @alice:example.com --no-encryption');
+      expect(
+          json.decode(
+              FakeMatrixApi.calledEndpoints['/client/r0/createRoom']?.first),
+          {'preset': 'private_chat'});
+    });
+
     test('discardsession', () async {
       if (olmEnabled) {
         await client.encryption?.keyManager.createOutboundGroupSession(room.id);
@@ -287,6 +309,11 @@ void main() {
                 null,
             false);
       }
+    });
+
+    test('create', () async {
+      await room.sendTextEvent('/clearcache');
+      expect(room.client.prevBatch, null);
     });
 
     test('dispose client', () async {